Add files via upload
Riemann Hypothesis: The Riemann hypothesis is a conjecture that the Riemann zeta function has its zeros only at the negative even integers and complex numbers with real part 1/2. Many consider it to be the most important unsolved problem in pure mathematics. The hypothesis indirectly inspires the distribution of primes. Thus here I have attempted to formulate a program that quantifies available primes before a given limit number. That is to say, it returns all primes less than the number the user inputs.