Add files via upload
Benford's Law: Benford's law, also called the Newcomb–Benford law is an observation that in many real-life sets of numerical data, the leading digit is likely to be small. In sets that obey the law, the number 1 appears as the leading significant digit about 30 % of the time, while 9 appears as the leading significant digit less than 5 % of the time. The numbers in a Collatz Conjecture sequence obey the law perfectly and are clearly visible for Collatz sequencing of large numbers.